feat: merge trip time and details components
