+++ frontend/src/components/BookingWizard/BookingWizard.tsx
@@ -1,66 +1,36 @@
-import { useState, useEffect } from 'react';
-import { Stepper, Step, StepLabel, Box } from '@mui/material';
-import SelectTimeStep from './SelectTimeStep';
-import TripDetailsStep from './TripDetailsStep';
+import { useState } from 'react';
+import { Box, Button } from '@mui/material';
+import TripDetails from './TripDetails';
 import PaymentStep from './PaymentStep';
 import { MapProvider } from '@/components/MapProvider';
 import { MapRoute } from '@/components/MapRoute';
 import { BookingFormData } from '@/types/BookingFormData';
-import { useAuth } from '@/contexts/AuthContext';
-
-const steps = ['Select time', 'Trip details', 'Payment'];
 
 export default function BookingWizard() {
-  const { userName, user, phone } = useAuth();
-  const [active, setActive] = useState(0);
   const [form, setForm] = useState<BookingFormData>({
     passengers: 1,
     notes: '',
-    customer: {
-      name: userName ?? undefined,
-      email: user?.email ?? undefined,
-      phone: phone ?? user?.phone ?? undefined,
-    },
     pickupValid: false,
     dropoffValid: false,
   });
-  useEffect(() => {
-    setForm((f) => ({
-      ...f,
-      customer: {
-        ...f.customer,
-        name: userName ?? undefined,
-        email: user?.email ?? undefined,
-        phone: phone ?? user?.phone ?? undefined,
-      },
-    }));
-  }, [userName, user?.email, phone, user?.phone]);
+  const [confirmed, setConfirmed] = useState(false);
   const update = (data: Partial<BookingFormData>) => {
     setForm((f) => ({ ...f, ...data }));
   };
-  const next = (data: Partial<BookingFormData>) => {
-    update(data);
-    setActive((s) => s + 1);
-  };
-  const back = () => setActive((s) => s - 1);
+
+  if (confirmed) {
+    return (
+      <PaymentStep
+        data={form as Required<BookingFormData>}
+        onBack={() => setConfirmed(false)}
+      />
+    );
+  }
 
   return (
     <Box>
-      <Stepper activeStep={active} sx={{ mb: 2 }}>
-        {steps.map((label) => (
-          <Step key={label}>
-            <StepLabel>{label}</StepLabel>
-          </Step>
-        ))}
-      </Stepper>
-      {active === 0 && <SelectTimeStep data={form} onNext={next} />}
-      {active === 1 && (
-        <TripDetailsStep data={form} onChange={update} onNext={next} onBack={back} />
-      )}
-      {active === 2 && (
-        <PaymentStep data={form as Required<BookingFormData>} onBack={back} />
-      )}
-      {active > 0 && form.pickupValid && form.dropoffValid && (
+      <TripDetails data={form} onChange={update} />
+      {form.pickupValid && form.dropoffValid && (
         <Box mt={2}>
           <MapProvider>
             <MapRoute
@@ -69,6 +39,13 @@ export default function BookingWizard() {
               rideTime={form.pickup_when}
             />
           </MapProvider>
+          <Button
+            sx={{ mt: 2 }}
+            variant="contained"
+            onClick={() => setConfirmed(true)}
+          >
+            Confirm booking
+          </Button>
         </Box>
       )}
     </Box>
